+++ pages/highscores/index.js
@@ -3,6 +3,15 @@ import { nanoid } from "nanoid";
 import { useState, useEffect } from "react";
 import styled from "styled-components";
 
+const StyledContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  max-height: 100vh;
+  width: 100vw;
+  overflow-y: scroll;
+`;
+
 const StyledHighscoresList = styled.ul`
   list-style: none;
   max-width: 440px;
@@ -10,16 +19,10 @@ const StyledHighscoresList = styled.ul`
   padding: 0;
 `;
 
-const StyledContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-`;
-
 const StyledListItem = styled.li`
   border: 2px solid black;
-  margin: 5px;
   padding: 7px 7px 7px 15px;
+  margin: 10px 0;
 `;
 
 export default function Highscores() {
